% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{column_mapper}
\alias{column_mapper}
\title{Convert the input from the dropdown menu from human readable name to column name.}
\usage{
column_mapper(column_name)
}
\arguments{
\item{column_name}{The name of the column to be mapped to the variable name.}
}
\value{
Returns the column name for the given variable name.
}
\description{
Convert the input from the dropdown menu from human readable name to column name.
}
