% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{comparison_graph}
\alias{comparison_graph}
\title{Plot the the two selected variables.}
\usage{
comparison_graph(
  dataset,
  x_comp,
  y_comp,
  comp_one_var,
  color_theme,
  graph_type,
  alpha_value
)
}
\arguments{
\item{dataset, }{name of the dataset description}

\item{x_comp, }{the variable that should be put on the x-axis.}

\item{y_comp, }{the variable that should be put on the y-axis.}
}
\value{
Returns a graph.
}
\description{
Plot the the two selected variables.
}
