% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{gender_dist}
\alias{gender_dist}
\title{Plot the gender distribution}
\usage{
gender_dist(dataset)
}
\arguments{
\item{dataset, }{name of the dataset description}
}
\value{
Bar graph.
}
\description{
Plot the gender distribution
}
