% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{load_dataset}
\alias{load_dataset}
\title{Import the dataset to be used.}
\usage{
load_dataset(dataset_path)
}
\arguments{
\item{dataset_path, }{Provide the path to the Lifelines dataset. description.}
}
\value{
A dataframe with the provided dataset.
}
\description{
Import the dataset to be used.
}
