% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{longer_df}
\alias{longer_df}
\title{Make the data "long" for use in the time graph.}
\usage{
longer_df(dataframe)
}
\arguments{
\item{dataframe}{The dataframe to be used.}
}
\value{
df returns the dataframe in long format.
}
\description{
Make the data "long" for use in the time graph.
}
