% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{make_factors_df}
\alias{make_factors_df}
\title{Internal function to make factors from some colums of the dataframe.}
\usage{
make_factors_df(dataframe)
}
\arguments{
\item{dataframe, }{load the dataframe to be used. description}
}
\description{
Internal function to make factors from some colums of the dataframe.
}
