% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{pregnancies_amount}
\alias{pregnancies_amount}
\title{Plot the amount of pregnancies per woman.}
\usage{
pregnancies_amount(dataset)
}
\arguments{
\item{dataset, }{name of the dataset description}
}
\value{
Returns a boxplot.
}
\description{
Plot the amount of pregnancies per woman.
}
