% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{progression_graph}
\alias{progression_graph}
\title{Plot the the two selected variables over time (T1, T2 and T3)}
\usage{
progression_graph(dataset, x_comp, y_comp, count_var, color_theme)
}
\arguments{
\item{dataset, }{name of the dataset description}

\item{x_comp, }{the variable that should be put on the x-axis.}

\item{y_comp, }{the variable that should be put on the y-axis.}
}
\value{
Returns a graph.
}
\description{
Plot the the two selected variables over time (T1, T2 and T3)
}
