% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{run_datadashboard}
\alias{run_datadashboard}
\title{Start the GUI}
\usage{
run_datadashboard(
  dataset_path = here("Data", "Lifelines Public Health dataset - 2024.csv")
)
}
\arguments{
\item{dataset_path, }{Provide the path to the Lifelines dataset. description.}
}
\value{
A dataframe with the provided dataset.
}
\description{
Start the GUI
}
