% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{sig_calculator}
\alias{sig_calculator}
\title{Calculate the significance for a given.}
\usage{
sig_calculator(dataset, column_name, given_value)
}
\arguments{
\item{dataset}{The dataset to be used.}

\item{column_name}{The columns to be used.}

\item{given_value}{The user specified value to use.}
}
\description{
Calculate the significance for a given.
}
