% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{weight_dist}
\alias{weight_dist}
\title{Plot the weight distribution for all the people.}
\usage{
weight_dist(dataset)
}
\arguments{
\item{dataset, }{name of the dataset description}
}
\value{
Returns a boxplot.
}
\description{
Plot the weight distribution for all the people.
}
