% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{finance_neighborhood_cor}
\alias{finance_neighborhood_cor}
\title{Plot the correlation between montly income and the satisfaction score.}
\usage{
finance_neighborhood_cor(dataset)
}
\arguments{
\item{dataset, }{name of the dataset description}
}
\value{
Jitter graph.
}
\description{
Plot the correlation between montly income and the satisfaction score.
}
